# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeilGraph",
    "description": "Local-first privacy compiler with Identity Exposure Graph, L1\u2013L5 transformations, fail-closed adversarial verification, cryptographic retention expiry, signed proof packages and Phase-2 production/security/scale controls.",
    "version": "0.6.1-phase2-final-hardening"
  },
  "paths": {
...
    "/api/v1/jobs": {
      "post": {
        "summary": "Create Job",
        "operationId": "create_job_api_v1_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/files": {
      "post": {
        "summary": "Upload File",
        "operationId": "upload_file_api_v1_jobs__job_id__files_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_api_v1_jobs__job_id__files_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/files/{file_id}/analyse": {
      "post": {
        "summary": "Analyse File",
        "operationId": "analyse_file_api_v1_jobs__job_id__files__file_id__analyse_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/files/{file_id}/entities": {
      "get": {
        "summary": "List Entities",
        "operationId": "list_entities_api_v1_jobs__job_id__files__file_id__entities_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityWithMentions"
                  },
                  "title": "Response List Entities Api V1 Jobs  Job Id  Files  File Id  Entities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/files/{file_id}/transform": {
      "post": {
        "summary": "Transform File",
        "operationId": "transform_file_api_v1_jobs__job_id__files__file_id__transform_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "File Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransformRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransformResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/jobs/{job_id}/destroy": {
      "delete": {
        "summary": "Destroy Job",
        "operationId": "destroy_job_api_v1_jobs__job_id__destroy_delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DestructionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "file_id": {
            "type": "string",
            "title": "File Id"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileType"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "scanned_pages": {
            "type": "integer",
            "title": "Scanned Pages"
          },
          "canonical_entities": {
            "type": "integer",
            "title": "Canonical Entities"
          },
          "mentions": {
            "type": "integer",
            "title": "Mentions"
          },
          "direct_identifier_mentions": {
            "type": "integer",
            "title": "Direct Identifier Mentions"
          },
          "quasi_identifier_mentions": {
            "type": "integer",
            "title": "Quasi Identifier Mentions"
          },
          "visual_mentions": {
            "type": "integer",
            "title": "Visual Mentions"
          },
          "pending_reviews": {
            "type": "integer",
            "title": "Pending Reviews"
          },
          "privacy_ir_schema": {
            "type": "string",
            "title": "Privacy Ir Schema"
          },
          "privacy_ir_units": {
            "type": "integer",
            "title": "Privacy Ir Units"
          },
          "privacy_ir_commitment_sha256": {
            "type": "string",
            "title": "Privacy Ir Commitment Sha256"
          },
          "structured_format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Format"
          },
          "structured_records": {
            "type": "integer",
            "title": "Structured Records",
            "default": 0
          },
          "structured_fields": {
            "type": "integer",
            "title": "Structured Fields",
            "default": 0
          },
          "structured_sheets": {
            "type": "integer",
            "title": "Structured Sheets",
            "default": 0
          },
          "structured_cells": {
            "type": "integer",
            "title": "Structured Cells",
            "default": 0
          },
          "structured_schema_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Structured Schema Sha256"
          },
          "docx_text_parts": {
            "type": "integer",
            "title": "Docx Text Parts",
            "default": 0
          },
          "docx_media_images": {
            "type": "integer",
            "title": "Docx Media Images",
            "default": 0
          },
          "docx_units": {
            "items": {
              "$ref": "#/components/schemas/DocxEvidenceUnitResponse"
            },
            "type": "array",
            "title": "Docx Units"
          },
          "video_duration_seconds": {
            "type": "number",
            "title": "Video Duration Seconds",
            "default": 0.0
          },
          "video_fps": {
            "type": "number",
            "title": "Video Fps",
            "default": 0.0
          },
          "video_width": {
            "type": "integer",
            "title": "Video Width",
            "default": 0
          },
          "video_height": {
            "type": "integer",
            "title": "Video Height",
            "default": 0
          },
          "video_total_frames": {
            "type": "integer",
            "title": "Video Total Frames",
            "default": 0
          },
          "video_sampled_frames": {
            "type": "integer",
            "title": "Video Sampled Frames",
            "default": 0
          },
          "video_security_frames_analyzed": {
            "type": "integer",
            "title": "Video Security Frames Analyzed",
            "default": 0
          },
          "video_security_detection_frames": {
            "type": "integer",
            "title": "Video Security Detection Frames",
            "default": 0
          },
          "video_novel_security_frames": {
            "type": "integer",
            "title": "Video Novel Security Frames",
            "default": 0
          },
          "video_security_coverage_percent": {
            "type": "number",
            "title": "Video Security Coverage Percent",
            "default": 0.0
          },
          "video_security_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Security Policy"
          },
          "video_has_audio": {
            "type": "boolean",
            "title": "Video Has Audio",
            "default": false
          },
          "video_audio_policy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Audio Policy"
          },
          "video_units": {
            "items": {
              "$ref": "#/components/schemas/VideoEvidenceUnitResponse"
            },
            "type": "array",
            "title": "Video Units"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "file_id",
          "file_type",
          "page_count",
          "scanned_pages",
          "canonical_entities",
          "mentions",
          "direct_identifier_mentions",
          "quasi_identifier_mentions",
          "visual_mentions",
          "pending_reviews",
          "privacy_ir_schema",
          "privacy_ir_units",
          "privacy_ir_commitment_sha256",
          "status"
        ],
        "title": "AnalysisResponse"
      },
      "AudienceProfile": {
        "type": "string",
        "enum": [
          "PUBLIC_RELEASE",
          "RESEARCH_PARTNER",
          "INTERNAL_OPERATIONS"
        ],
        "title": "AudienceProfile"
      },
...
      "Body_upload_file_api_v1_jobs__job_id__files_post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_file_api_v1_jobs__job_id__files_post"
      },
...
      "EntityWithMentions": {
        "properties": {
          "entity": {
            "$ref": "#/components/schemas/CanonicalEntityResponse"
          },
          "mentions": {
            "items": {
              "$ref": "#/components/schemas/EntityMentionResponse"
            },
            "type": "array",
            "title": "Mentions"
          }
        },
        "type": "object",
        "required": [
          "entity",
          "mentions"
        ],
        "title": "EntityWithMentions"
      },
...
      "FileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "original_filename": {
            "type": "string",
            "title": "Original Filename"
          },
          "file_type": {
            "$ref": "#/components/schemas/FileType"
          },
          "media_type": {
            "type": "string",
            "title": "Media Type"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "original_filename",
          "file_type",
          "media_type",
          "sha256",
          "page_count",
          "status",
          "created_at"
        ],
        "title": "FileResponse"
      },
      "FileType": {
        "type": "string",
        "enum": [
          "PDF",
          "IMAGE",
          "TEXT",
          "DATASET",
          "DOCX",
          "VIDEO"
        ],
        "title": "FileType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobCreate": {
        "properties": {
          "purpose": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Purpose"
          },
          "recipient": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Recipient"
          },
          "audience_profile": {
            "$ref": "#/components/schemas/AudienceProfile",
            "default": "PUBLIC_RELEASE"
          },
          "privacy_level": {
            "$ref": "#/components/schemas/PrivacyLevel",
            "default": 3
          },
          "retention_seconds": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Retention Seconds",
            "default": 3600
          }
        },
        "type": "object",
        "required": [
          "purpose",
          "recipient"
        ],
        "title": "JobCreate"
      },
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "purpose": {
            "type": "string",
            "title": "Purpose"
          },
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "audience_profile": {
            "$ref": "#/components/schemas/AudienceProfile"
          },
          "privacy_level": {
            "$ref": "#/components/schemas/PrivacyLevel"
          },
          "retention_seconds": {
            "type": "integer",
            "title": "Retention Seconds"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "purpose",
          "recipient",
          "audience_profile",
          "privacy_level",
          "retention_seconds",
          "expires_at",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "JobResponse"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "CREATED",
          "UPLOADED",
          "ANALYSED",
          "HUMAN_REVIEW_REQUIRED",
          "TRANSFORMED",
          "VERIFIED",
          "BLOCKED",
          "DESTROYED",
          "FAILED"
        ],
        "title": "JobStatus"
      },
...
      "PrivacyLevel": {
        "type": "integer",
        "enum": [
          1,
          2,
          3,
          4,
          5
        ],
        "title": "PrivacyLevel"
      },
...
      "TransformRequest": {
        "properties": {
          "privacy_level": {
            "$ref": "#/components/schemas/PrivacyLevel",
            "default": 3
          }
        },
        "type": "object",
        "title": "TransformRequest"
      },
...
    }
  }
}